engage_scheduler/requirements.txt
-i https://pypi.python.org/simple
beautifulsoup4==4.8.2
boto3==1.12.13
botocore==1.15.13
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
docutils==0.15.2
engage-scraper==0.0.37
future==0.18.2
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
idna==2.9
jmespath==0.9.5
jsonpickle==1.3
mailjet-rest==1.3.3
oauthlib==3.1.0
pillow==7.0.0
protobuf==3.11.3
psycopg2-binary==2.8.4
pyasn1-modules==0.2.8
pyasn1==0.4.8
python-dateutil==2.8.1
python-twitter==3.5
pytz==2019.3
requests-oauthlib==1.3.0
requests==2.23.0
rsa==4.0
s3transfer==0.3.3
schedule==0.6.0
simple-latex==0.0.22
six==1.14.0
soupsieve==2.0
sqlalchemy==1.3.13
urllib3==1.25.8 ; python_version != '3.4'
